Handling wrong pageId and languageId
In case pageId in the URL points to a non-exisiting page, the
framework will assume the first page from the site list.
Also if languageId points to a non-existing language, the
framework will assume the first language from the list.

The alternative solution in case of invalid page/language values
would be to return nice error message, but I decided to go for a
fallback to defaults, because I saw that this solution has more
value for the user than the error message.